fix: adjust routing structure and rename Home/Landing pages
