+++ app/src/tests/placeholder.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@test-utils';
-import HomePage from '@/pages/Home.page';
+import HomePage from '@/pages/Dashboard.page';
 
 describe('HomePage component', () => {
   it('exists', () => {
